powerpc/vsx: implement vec_sld for vector_double
